+++ src/di.rs
@@ -19,30 +19,11 @@ use std::{
     collections::{HashMap, HashSet},
     fmt::{Debug, Formatter, Write},
     future::Future,
-    ops::Deref,
     panic::Location,
     sync::Arc,
 };
 
 use crate::Type;
-
-/// A DI container from which we can extract a value of a given type.
-///
-/// There are two possible ways to handle the situation when your container
-/// cannot return a value of specified type:
-///
-/// 1. Do not implement [`DependencySupplier`] for the type. It often requires
-/// some type-level manipulations.
-/// 2. Runtime panic. Be careful in this case: check whether you add your type
-/// to the container.
-///
-/// A concrete solution is left to a particular implementation.
-pub trait DependencySupplier<Value> {
-    /// Get the value.
-    ///
-    /// We assume that all values are stored in `Arc<_>`.
-    fn get(&self) -> Arc<Value>;
-}
 
 /// A DI container with multiple dependencies.
 ///
@@ -56,7 +37,7 @@ pub trait DependencySupplier<Value> {
 ///
 /// ```
 /// # use std::sync::Arc;
-/// use dptree::di::{DependencyMap, DependencySupplier};
+/// use dptree::di::DependencyMap;
 ///
 /// let mut container = DependencyMap::new();
 /// container.insert(5_i32);
@@ -76,7 +57,7 @@ pub trait DependencySupplier<Value> {
 ///
 /// ```should_panic
 /// # use std::sync::Arc;
-/// use dptree::di::{DependencyMap, DependencySupplier};
+/// use dptree::di::DependencyMap;
 /// let mut container = DependencyMap::new();
 /// container.insert(10i32);
 /// container.insert(true);
@@ -142,6 +123,30 @@ impl DependencyMap {
             .map(|dep| dep.inner.downcast().expect("Values are stored by TypeId"))
     }
 
+    /// Retrieves the value of type `V` from this container.
+    ///
+    /// ## Panics
+    ///
+    /// If the container has no value of type `V`.
+    pub fn get<V>(&self) -> Arc<V>
+    where
+        V: Send + Sync + 'static,
+    {
+        self.map
+            .get(&TypeId::of::<V>())
+            .unwrap_or_else(|| {
+                panic!(
+                    "{} was requested, but not provided. Available types:\n{}",
+                    std::any::type_name::<V>(),
+                    self.available_types()
+                )
+            })
+            .clone()
+            .inner
+            .downcast::<V>()
+            .expect("Checked by .unwrap_or_else()")
+    }
+
     fn available_types(&self) -> String {
         let mut list = String::new();
 
@@ -159,49 +164,15 @@ impl Debug for DependencyMap {
     }
 }
 
-impl<V> DependencySupplier<V> for DependencyMap
-where
-    V: Send + Sync + 'static,
-{
-    fn get(&self) -> Arc<V> {
-        self.map
-            .get(&TypeId::of::<V>())
-            .unwrap_or_else(|| {
-                panic!(
-                    "{} was requested, but not provided. Available types:\n{}",
-                    std::any::type_name::<V>(),
-                    self.available_types()
-                )
-            })
-            .clone()
-            .inner
-            .downcast::<V>()
-            .expect("Checked by .unwrap_or_else()")
-    }
-}
-
-impl<V, S> DependencySupplier<V> for Arc<S>
-where
-    S: DependencySupplier<V>,
-{
-    fn get(&self) -> Arc<V> {
-        self.deref().get()
-    }
-}
-
 /// Converts functions into [`CompiledFn`].
 ///
-/// The function must follow some rules, to be usable with DI:
-///
-/// 1. For each function parameter of type `T`, `Input` must satisfy
-/// `DependencySupplier<T>`.
-/// 2. The function must be of 0-9 arguments.
-/// 3. The function must return [`Future`].
-pub trait Injectable<Input, Output, FnArgs>
+/// For a function to be convertible into [`CompiledFn`], it must be of 0-9
+/// arguments and return [`Future`].
+pub trait Injectable<Output, FnArgs>
 where
     Output: 'static,
 {
-    fn inject<'a>(&'a self, container: &'a Input) -> CompiledFn<'a, Output>;
+    fn inject<'a>(&'a self, container: &'a DependencyMap) -> CompiledFn<'a, Output>;
 
     /// Returns the set of types that this function depends on. Used only for
     /// run-time "type checking".
@@ -228,10 +199,8 @@ pub struct Asyncify<F>(pub F);
 
 macro_rules! impl_into_di {
     ($($generic:ident),*) => {
-        impl<Func, Input, Output, Fut, $($generic),*> Injectable<Input, Output, ($($generic,)*)> for Func
+        impl<Func, Output, Fut, $($generic),*> Injectable<Output, ($($generic,)*)> for Func
         where
-            Input: $(DependencySupplier<$generic> +)*,
-            Input: Send + Sync,
             Func: Fn($($generic),*) -> Fut + Send + Sync + 'static,
             Fut: Future<Output = Output> + Send + 'static,
             Output: 'static,
@@ -239,7 +208,7 @@ macro_rules! impl_into_di {
         {
             #[allow(non_snake_case)]
             #[allow(unused_variables)]
-            fn inject<'a>(&'a self, container: &'a Input) -> CompiledFn<'a, Output> {
+            fn inject<'a>(&'a self, container: &'a DependencyMap) -> CompiledFn<'a, Output> {
                 Arc::new(move || {
                     $(let $generic = std::borrow::Borrow::<$generic>::borrow(&container.get()).clone();)*
                     let fut = self( $( $generic ),* );
@@ -254,17 +223,15 @@ macro_rules! impl_into_di {
             }
         }
 
-        impl<Func, Input, Output, $($generic),*> Injectable<Input, Output, ($($generic,)*)> for Asyncify<Func>
+        impl<Func, Output, $($generic),*> Injectable<Output, ($($generic,)*)> for Asyncify<Func>
         where
-            Input: $(DependencySupplier<$generic> +)*,
-            Input: Send + Sync,
             Func: Fn($($generic),*) -> Output + Send + Sync + 'static,
             Output: Send + 'static,
             $($generic: Clone + Send + Sync + 'static),*
         {
             #[allow(non_snake_case)]
             #[allow(unused_variables)]
-            fn inject<'a>(&'a self, container: &'a Input) -> CompiledFn<'a, Output> {
+            fn inject<'a>(&'a self, container: &'a DependencyMap) -> CompiledFn<'a, Output> {
                 let Asyncify(this) = self;
                 Arc::new(move || {
                     $(let $generic = std::borrow::Borrow::<$generic>::borrow(&container.get()).clone();)*
@@ -298,7 +265,7 @@ impl_into_di!(A, B, C, D, E, F, G, H, I);
 /// # Examples
 ///
 /// ```
-/// use dptree::di::{DependencyMap, DependencySupplier};
+/// use dptree::di::DependencyMap;
 ///
 /// let map = dptree::deps![123, "abc", true];
 ///
@@ -320,18 +287,6 @@ macro_rules! deps {
             $(map.insert($dep);)*
             map
         }
-    }
-}
-
-/// Insert some value to a container.
-pub trait Insert<Value> {
-    /// Inserts `value` into itself, returning the previous value, if exists.
-    fn insert(&mut self, value: Value) -> Option<Arc<Value>>;
-}
-
-impl<T: Send + Sync + 'static> Insert<T> for DependencyMap {
-    fn insert(&mut self, value: T) -> Option<Arc<T>> {
-        DependencyMap::insert(self, value)
     }
 }